cmd_vel_mux: Fix assert failure on startup
Fix a startup race wherein cmd_vel_mux could attempt to publish to
the /active publisher before that object is created. Depending on what
random bits happen to be in RAM at that time, this could either quietly
miss the initial "idle" publish or it could cause an ASSERT abort due to
attempt to publish to an invalid publisher object.